feat: thin strategy layer
