+++ simaple/simulate/strategy/base.py
@@ -1,79 +1,16 @@
-import functools
-from typing import Callable, Generator, Optional, cast
+from typing import Callable, Generator, Optional
 
-from simaple.simulate.base import Action, BehaviorGenerator, Event, ViewerType
+from simaple.simulate.base import Event, ViewerType
+from simaple.simulate.engine import OperationEngineProtocol
 from simaple.simulate.policy.base import Operation
 
-ActionGeneratorType = Generator[Action, list[Event], None]
+RuntimeContextType = tuple[ViewerType, list[Event]]
+
+OperationGenerator = Generator[Operation, RuntimeContextType, RuntimeContextType]
+OperationGeneratorProto = Callable[[RuntimeContextType], OperationGenerator]
 
 
-class BehaviorStrategy:
-    """Generate Exec Op Runtime, and provides a generator interface to handle events.
-    use case:
-    >>> def exec_use(op, event: list[Event]) -> ExecOpType:
-    >>>     ...
-    >>> runtime = BehaviorStrategy(exec_use)
-    >>> for behavior in runtime.handle(op):
-    >>>     action = behavior(events)
-    >>>     events = engine.play(action)
-
-    """
-
-    def __init__(
-        self, gen_method: Callable[[Operation, list[Event]], ActionGeneratorType]
-    ):
-        self._gen_method = gen_method
-        self._gen: Optional[ActionGeneratorType] = None
-        self._end: bool = False
-
-    def handle(
-        self,
-        op: Operation,
-    ) -> BehaviorGenerator:
-        self._op: Operation = op
-        while not self._end:
-            yield self
-
-    def __call__(self, events: list[Event]) -> Optional[Action]:
-        if self._gen is None:
-            self._gen = self._gen_method(self._op, events)
-            return next(self._gen)
-
-        try:
-            return self._gen.send(events)
-        except StopIteration as e:
-            self._end = True
-            return cast(Optional[Action], e.value)
-
-    @classmethod
-    def operation_handler(
-        cls, gen_method: Callable[[Operation, list[Event]], ActionGeneratorType]
-    ):
-        """Wraps Exec Op Runtime, and provides a generator interface to handle events.
-        use case:
-        >>> @BehaviorGenerator.operation_handler
-        >>> def exec_use(op, event: list[Event]) -> ExecOpType:
-        >>>     ...
-        >>> for behavior in exec_use(op):
-        >>>     action = behavior(events)
-        >>>     events = engine.play(action)
-        """
-
-        @functools.wraps(gen_method)
-        def _wrapper(op: Operation):
-            gen = cls(gen_method)
-            return gen.handle(op)
-
-        return _wrapper
-
-
-PolicyContextType = tuple[ViewerType, list[Event]]
-
-OperationGenerator = Generator[Operation, PolicyContextType, PolicyContextType]
-OperationGeneratorProto = Callable[[PolicyContextType], OperationGenerator]
-
-
-PolicyType = Callable[[PolicyContextType], list[Operation]]
+StrategyType = Callable[[RuntimeContextType], list[Operation]]
 
 
 class PolicyWrapper:
@@ -81,9 +18,17 @@ class PolicyWrapper:
         self.operation_generator_proto = operation_generator_proto
         self._operation_generator: Optional[OperationGenerator] = None
 
-    def __call__(self, context: PolicyContextType) -> Operation:
+    def __call__(self, context: RuntimeContextType) -> Operation:
         if self._operation_generator is None:
             self._operation_generator = self.operation_generator_proto(context)
             return next(self._operation_generator)
 
         return self._operation_generator.send(context)
+
+
+def exec_by_strategy(
+    engine: OperationEngineProtocol, policy: StrategyType, early_stop: int = -1
+) -> None:
+    operations = policy((engine.get_current_viewer(), engine.get_buffered_events()))
+    for op in operations:
+        engine.exec(op, early_stop=early_stop)
