+++ test_data/test_container_runtime.py
@@ -3,6 +3,7 @@ import pytest
 import simaple.simulate.component.skill  # pylint: disable=W0611
 from simaple.container.simulation import SimulationContainer
 from test_data.target import get_test_settings
+from simaple.simulate.strategy.base import exec_by_strategy
 
 
 @pytest.mark.parametrize("setting, jobtype, expected", get_test_settings())
@@ -17,7 +18,7 @@ def test_actor(setting, jobtype, expected):
     policy = container.skill_profile().get_default_policy()
 
     while engine.get_current_viewer()("clock") < 50_000:
-        engine.exec_policy(policy, early_stop=50_000)
+        exec_by_strategy(engine, policy, early_stop=50_000)
 
     report = engine.create_full_report()
